requirements.txt
streamlit>=1.30.0
requests>=2.31.0
lxml>=5.1.0
pandas>=2.1.0
scikit-learn>=1.3.0
nltk>=3.8.0
beautifulsoup4>=4.12.0
plotly>=5.18.0
python-dateutil>=2.8.0
chardet>=5.2.0
openpyxl>=3.1.0
curl_cffi>=0.15.0
gunicorn>=21.2.0
watchdog>=4.0.0
